(change)Employee: moved employee to model folder
